# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Naxchivan Residence ERP API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ApartmentStatus =
  (typeof ApartmentStatus)[keyof typeof ApartmentStatus];

export const ApartmentStatus = {
  available: "available",
  sold: "sold",
  reserved: "reserved",
} as const;

export type ObjectType = (typeof ObjectType)[keyof typeof ObjectType];

export const ObjectType = {
  object: "object",
  garage: "garage",
} as const;

export type ObjectStatus = (typeof ObjectStatus)[keyof typeof ObjectStatus];

export const ObjectStatus = {
  available: "available",
  sold: "sold",
  rented: "rented",
} as const;

export type SaleType = (typeof SaleType)[keyof typeof SaleType];

export const SaleType = {
  cash: "cash",
  credit: "credit",
} as const;

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus];

export const PaymentStatus = {
  pending: "pending",
  paid: "paid",
  overdue: "overdue",
} as const;

export type RentalStatus = (typeof RentalStatus)[keyof typeof RentalStatus];

export const RentalStatus = {
  active: "active",
  ended: "ended",
} as const;

export type InternetStatus =
  (typeof InternetStatus)[keyof typeof InternetStatus];

export const InternetStatus = {
  active: "active",
  expired: "expired",
  cancelled: "cancelled",
} as const;

export interface Block {
  id: number;
  name: string;
  apartmentCount: number;
}

export interface CreateBlockInput {
  name: string;
}

export interface Apartment {
  id: number;
  blockId: number;
  blockName: string;
  number: string;
  floor: number;
  area: number;
  status: ApartmentStatus;
  pricePerSqm?: number;
  totalPrice?: number;
}

export interface CreateApartmentInput {
  blockId: number;
  number: string;
  floor: number;
  area: number;
}

export interface ObjectGarage {
  id: number;
  number: string;
  area: number;
  type: ObjectType;
  status: ObjectStatus;
  pricePerSqm?: number;
  totalPrice?: number;
}

export interface CreateObjectInput {
  number: string;
  area: number;
  type: ObjectType;
}

export interface Customer {
  id: number;
  firstName: string;
  lastName: string;
  fin?: string;
  phone: string;
  address?: string;
}

export interface CreateCustomerInput {
  firstName: string;
  lastName: string;
  fin?: string;
  phone: string;
  address?: string;
}

export type SaleAssetType = (typeof SaleAssetType)[keyof typeof SaleAssetType];

export const SaleAssetType = {
  apartment: "apartment",
  object: "object",
  garage: "garage",
} as const;

export interface Sale {
  id: number;
  customerId: number;
  customerName: string;
  assetType: SaleAssetType;
  assetId: number;
  assetDescription: string;
  saleType: SaleType;
  totalAmount: number;
  downPayment: number;
  installmentMonths: number;
  monthlyPayment: number;
  paidAmount: number;
  remainingAmount: number;
  saleDate: string;
  progressPercent: number;
}

export type CustomerDetail = Customer & {
  sales: Sale[];
};

export interface Installment {
  id: number;
  saleId: number;
  dueDate: string;
  amount: number;
  paidDate?: string | null;
  status: PaymentStatus;
  installmentNumber: number;
}

export type SaleDetail = Sale & {
  installments: Installment[];
};

export type CreateSaleInputAssetType =
  (typeof CreateSaleInputAssetType)[keyof typeof CreateSaleInputAssetType];

export const CreateSaleInputAssetType = {
  apartment: "apartment",
  object: "object",
  garage: "garage",
} as const;

export interface CreateSaleInput {
  customerId: number;
  assetType: CreateSaleInputAssetType;
  assetId: number;
  saleType: SaleType;
  downPayment: number;
  installmentMonths?: number;
  pricePerSqm: number;
}

export interface CalculateInput {
  area: number;
  pricePerSqm: number;
  downPayment: number;
  installmentMonths: number;
}

export interface CalculationResult {
  totalAmount: number;
  downPayment: number;
  creditAmount: number;
  monthlyPayment: number;
  installmentMonths: number;
}

export interface PayInstallmentInput {
  paidDate: string;
}

export type RentalAssetType =
  (typeof RentalAssetType)[keyof typeof RentalAssetType];

export const RentalAssetType = {
  object: "object",
  garage: "garage",
} as const;

export interface Rental {
  id: number;
  assetType: RentalAssetType;
  assetId: number;
  assetDescription: string;
  customerId: number;
  customerName: string;
  startDate: string;
  endDate: string;
  monthlyAmount: number;
  status: RentalStatus;
  daysRemaining: number;
}

export type CreateRentalInputAssetType =
  (typeof CreateRentalInputAssetType)[keyof typeof CreateRentalInputAssetType];

export const CreateRentalInputAssetType = {
  object: "object",
  garage: "garage",
} as const;

export interface CreateRentalInput {
  assetType: CreateRentalInputAssetType;
  assetId: number;
  customerId: number;
  startDate: string;
  endDate: string;
  monthlyAmount: number;
}

export interface UpdateRentalInput {
  status?: RentalStatus;
  endDate?: string;
}

export interface CommunalBill {
  id: number;
  assetType: string;
  assetId: number;
  assetDescription: string;
  ownerName: string;
  month: number;
  year: number;
  area: number;
  tariff: number;
  amount: number;
  status: PaymentStatus;
  paidDate?: string | null;
}

export interface GenerateBillsInput {
  month: number;
  year: number;
}

export interface InternetSubscription {
  id: number;
  assetType: string;
  assetId: number;
  assetDescription: string;
  ownerName: string;
  packageName: string;
  monthlyPrice: number;
  startDate: string;
  expiryDate: string;
  status: InternetStatus;
  daysRemaining: number;
}

export type CreateInternetInputAssetType =
  (typeof CreateInternetInputAssetType)[keyof typeof CreateInternetInputAssetType];

export const CreateInternetInputAssetType = {
  apartment: "apartment",
  object: "object",
  garage: "garage",
} as const;

export interface CreateInternetInput {
  assetType: CreateInternetInputAssetType;
  assetId: number;
  customerId: number;
  packageName: string;
  monthlyPrice: number;
  startDate: string;
}

export interface TariffsResponse {
  communalTariff: number;
  apartmentPricePerSqm: number;
  objectPricePerSqm: number;
  garagePricePerSqm: number;
}

export interface UpdateTariffsInput {
  communalTariff?: number;
  apartmentPricePerSqm?: number;
  objectPricePerSqm?: number;
  garagePricePerSqm?: number;
}

export interface DashboardStats {
  totalApartments: number;
  soldApartments: number;
  availableApartments: number;
  reservedApartments: number;
  totalObjects: number;
  availableObjects: number;
  totalGarages: number;
  availableGarages: number;
  totalSales: number;
  cashSales: number;
  creditSales: number;
  totalRevenue: number;
  pendingInstallments: number;
  overdueInstallments: number;
  activeRentals: number;
  pendingCommunalBills: number;
  activeInternetSubscriptions: number;
  totalCustomers: number;
  monthlyInstallmentIncome: number;
  monthlyCommunalIncome: number;
}

export type ListApartmentsParams = {
  blockId?: number;
  status?: ApartmentStatus;
};

export type ListObjectsParams = {
  type?: ObjectType;
  status?: ObjectStatus;
};

export type ListCustomersParams = {
  search?: string;
};

export type ListSalesParams = {
  customerId?: number;
  saleType?: SaleType;
};

export type ListInstallmentsParams = {
  saleId?: number;
  status?: PaymentStatus;
  overdue?: boolean;
};

export type ListRentalsParams = {
  status?: RentalStatus;
};

export type ListCommunalBillsParams = {
  month?: number;
  year?: number;
  status?: PaymentStatus;
};

export type GenerateCommunalBills201 = {
  count: number;
  bills: CommunalBill[];
};

export type ListInternetSubscriptionsParams = {
  status?: InternetStatus;
};