# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Naxchivan Residence ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /objects:
    get:
      operationId: listObjects
      tags: [objects]
      summary: List all objects and garages
      parameters:
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/ObjectType"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ObjectStatus"
      responses:
        "200":
          description: List of objects/garages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ObjectGarage"
    post:
      operationId: createObject
      tags: [objects]
      summary: Create an object or garage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateObjectInput"
      responses:
        "201":
          description: Created object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectGarage"
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List all sales
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: saleType
          in: query
          schema:
            $ref: "#/components/schemas/SaleType"
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create a sale and generate installment schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleInput"
      responses:
        "201":
          description: Created sale with installment schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleDetail"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Get sale details with installments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleDetail"
  /sales/calculate:
    post:
      operationId: calculateInstallment
      tags: [sales]
      summary: Calculate monthly installment amount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalculateInput"
      responses:
        "200":
          description: Calculation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalculationResult"
...
  /rentals:
    get:
      operationId: listRentals
      tags: [rentals]
      summary: List all rentals
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/RentalStatus"
      responses:
        "200":
          description: List of rentals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rental"
    post:
      operationId: createRental
      tags: [rentals]
      summary: Create a rental agreement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRentalInput"
      responses:
        "201":
          description: Created rental
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rental"
  /rentals/{id}:
    put:
      operationId: updateRental
      tags: [rentals]
      summary: Update rental (end/terminate)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRentalInput"
      responses:
        "200":
          description: Updated rental
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rental"
...
components:
  schemas:
...
    ObjectType:
      type: string
      enum: [object, garage]
    ObjectStatus:
      type: string
      enum: [available, sold, rented]
    SaleType:
      type: string
      enum: [cash, credit]
    PaymentStatus:
      type: string
      enum: [pending, paid, overdue]
    RentalStatus:
      type: string
      enum: [active, ended]
...
    ObjectGarage:
      type: object
      properties:
        id:
          type: integer
        number:
          type: string
        area:
          type: number
        type:
          $ref: "#/components/schemas/ObjectType"
        status:
          $ref: "#/components/schemas/ObjectStatus"
        pricePerSqm:
          type: number
        totalPrice:
          type: number
      required:
        - id
        - number
        - area
        - type
        - status
    CreateObjectInput:
      type: object
      properties:
        number:
          type: string
        area:
          type: number
        type:
          $ref: "#/components/schemas/ObjectType"
      required:
        - number
        - area
        - type
...
    Sale:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        assetType:
          type: string
          enum: [apartment, object, garage]
        assetId:
          type: integer
        assetDescription:
          type: string
        saleType:
          $ref: "#/components/schemas/SaleType"
        totalAmount:
          type: number
        downPayment:
          type: number
        installmentMonths:
          type: integer
        monthlyPayment:
          type: number
        paidAmount:
          type: number
        remainingAmount:
          type: number
        saleDate:
          type: string
        progressPercent:
          type: number
      required:
        - id
        - customerId
        - customerName
        - assetType
        - assetId
        - assetDescription
        - saleType
        - totalAmount
        - downPayment
        - installmentMonths
        - monthlyPayment
        - paidAmount
        - remainingAmount
        - saleDate
        - progressPercent
    SaleDetail:
      allOf:
        - $ref: "#/components/schemas/Sale"
        - type: object
          properties:
            installments:
              type: array
              items:
                $ref: "#/components/schemas/Installment"
          required:
            - installments
    CreateSaleInput:
      type: object
      properties:
        customerId:
          type: integer
        assetType:
          type: string
          enum: [apartment, object, garage]
        assetId:
          type: integer
        saleType:
          $ref: "#/components/schemas/SaleType"
        downPayment:
          type: number
        installmentMonths:
          type: integer
        pricePerSqm:
          type: number
      required:
        - customerId
        - assetType
        - assetId
        - saleType
        - downPayment
        - pricePerSqm
    CalculateInput:
      type: object
      properties:
        area:
          type: number
        pricePerSqm:
          type: number
        downPayment:
          type: number
        installmentMonths:
          type: integer
      required:
        - area
        - pricePerSqm
        - downPayment
        - installmentMonths
    CalculationResult:
      type: object
      properties:
        totalAmount:
          type: number
        downPayment:
          type: number
        creditAmount:
          type: number
        monthlyPayment:
          type: number
        installmentMonths:
          type: integer
      required:
        - totalAmount
        - downPayment
        - creditAmount
        - monthlyPayment
        - installmentMonths
    Installment:
      type: object
      properties:
        id:
          type: integer
        saleId:
          type: integer
        dueDate:
          type: string
        amount:
          type: number
        paidDate:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/PaymentStatus"
        installmentNumber:
          type: integer
      required:
        - id
        - saleId
        - dueDate
        - amount
        - status
        - installmentNumber
...
    Rental:
      type: object
      properties:
        id:
          type: integer
        assetType:
          type: string
          enum: [object, garage]
        assetId:
          type: integer
        assetDescription:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        monthlyAmount:
          type: number
        status:
          $ref: "#/components/schemas/RentalStatus"
        daysRemaining:
          type: integer
      required:
        - id
        - assetType
        - assetId
        - assetDescription
        - customerId
        - customerName
        - startDate
        - endDate
        - monthlyAmount
        - status
        - daysRemaining
    CreateRentalInput:
      type: object
      properties:
        assetType:
          type: string
          enum: [object, garage]
        assetId:
          type: integer
        customerId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        monthlyAmount:
          type: number
      required:
        - assetType
        - assetId
        - customerId
        - startDate
        - endDate
        - monthlyAmount
    UpdateRentalInput:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/RentalStatus"
        endDate:
          type: string
...