# lib/api-zod/src/generated/types/createRentalInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Naxchivan Residence ERP API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateRentalInputAssetType } from "./createRentalInputAssetType";

export interface CreateRentalInput {
  assetType: CreateRentalInputAssetType;
  assetId: number;
  customerId: number;
  startDate: string;
  endDate: string;
  monthlyAmount: number;
}