# lib/api-zod/src/generated/types/objectStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Naxchivan Residence ERP API
 * OpenAPI spec version: 0.1.0
 */

export type ObjectStatus = (typeof ObjectStatus)[keyof typeof ObjectStatus];

export const ObjectStatus = {
  available: "available",
  sold: "sold",
  rented: "rented",
} as const;