# lib/api-zod/src/generated/types/rental.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Naxchivan Residence ERP API
 * OpenAPI spec version: 0.1.0
 */
import type { RentalAssetType } from "./rentalAssetType";
import type { RentalStatus } from "./rentalStatus";

export interface Rental {
  id: number;
  assetType: RentalAssetType;
  assetId: number;
  assetDescription: string;
  customerId: number;
  customerName: string;
  startDate: string;
  endDate: string;
  monthlyAmount: number;
  status: RentalStatus;
  daysRemaining: number;
}