# lib/api-zod/src/generated/types/rentalStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Naxchivan Residence ERP API
 * OpenAPI spec version: 0.1.0
 */

export type RentalStatus = (typeof RentalStatus)[keyof typeof RentalStatus];

export const RentalStatus = {
  active: "active",
  ended: "ended",
} as const;